pyproject.toml
[project]
name = "multi-bank-fraud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker>=38.2.0",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
]
